bug render_with_signed_in_user

+++ app/controllers/dashboard/posts_controller.rb
@@ -133,33 +133,40 @@ class Dashboard::PostsController < ApplicationController
   def like 
     like = current_user.get_likeable @post 
     if like 
-      notification = Notification.find_by notifiable: like
-      ActionCable.server.broadcast "notifications:#{@post.user.id}", {action:'remove', notification:notification} 
+      if @post.user.id != current_user.id
+        notification = Notification.find_by notifiable: like
+        ActionCable.server.broadcast "notifications:#{@post.user.id}", {
+          action:'remove', 
+          notification:notification
+        } 
+        notification.destroy 
+      end
       like.destroy
-      notification.destroy 
     else 
       created_like = current_user.likes.create(likeable: @post)
-      notification = Notification.create(
-        notifiable: created_like,
-        message:"like your post: #{@post.title}", 
-        link: post_path(@post),
-        sender: current_user,
-        recipient: @post.user 
-      )
-      html_header = ApplicationController.render(
-        partial: 'shared/notification_item',
-        locals: { notification: notification }
-      )
-      html_toast = ApplicationController.render(
-        partial: 'shared/notification_toast',
-        locals: {notification: notification}
-      )
-      ActionCable.server.broadcast "notifications:#{@post.user.id}", {
-        action:'add', 
-        html_header:html_header, 
-        html_toast:html_toast,
-        notification:notification
-      } 
+      if @post.user.id != current_user.id
+        notification = Notification.create(
+          notifiable: created_like,
+          message:"like your post: #{@post.title}", 
+          link: post_path(@post),
+          sender: current_user,
+          recipient: @post.user 
+        )
+        html_header = ApplicationController.render(
+          partial: 'shared/notification_item',
+          locals: { notification: notification }
+        )
+        html_toast = ApplicationController.render(
+          partial: 'shared/notification_toast',
+          locals: {notification: notification}
+        )
+        ActionCable.server.broadcast "notifications:#{@post.user.id}", {
+          action:'add', 
+          html_header:html_header, 
+          html_toast:html_toast,
+          notification:notification
+        } 
+      end
     end
     render partial:"dashboard/posts/js_erb/like.js.erb" 
   end
